Allow noise_size == 0 without preventing ep calculation
Gaussian blurring can be done before invoking trackpy so a noise_size of zero should be able to be set. However, in previous setup this meant that either preprocess had to be turned off (no bandpassed image, no ep calculation), or that noise_size had to be set to zero (no ep calculation anymore). This fixes it.